Refactor SearchRestaurant to Restautants
