+++ data/src/main/java/com/app/data/entity/mapper/RestaurantEntityDataMapper.java
@@ -1,6 +1,6 @@
 package com.app.data.entity.mapper;
 
-import com.app.data.entity.restaurantSearch.RestaurantSearchEntity;
+import com.app.data.entity.restaurants.RestaurantEntity;
 import com.app.domain.model.restaurants.RestaurantsDataModel;
 
 import javax.inject.Inject;
@@ -10,7 +10,7 @@ public class RestaurantEntityDataMapper {
     @Inject public RestaurantEntityDataMapper() {
     }
 
-    public RestaurantsDataModel map(RestaurantSearchEntity restaurantSearchEntity) {
+    public RestaurantsDataModel map(RestaurantEntity restaurantEntity) {
         RestaurantsDataModel dataModel = new RestaurantsDataModel();
         return dataModel;
     }
