+++ data/src/main/java/com/app/data/entity/restaurants/Restaurant.java
@@ -1,5 +1,5 @@
 
-package com.app.data.entity.restaurantSearch;
+package com.app.data.entity.restaurants;
 
 import java.util.List;
 import com.google.gson.annotations.Expose;
