+++ data/src/main/java/com/app/data/entity/restaurants/User.java
@@ -1,10 +1,10 @@
 
-package com.app.data.entity.restaurantSearch;
+package com.app.data.entity.restaurants;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
-public class User_ {
+public class User {
 
     @SerializedName("name")
     @Expose
