+++ data/src/main/java/com/app/data/net/services/RestaurantsApi.java
@@ -1,6 +1,6 @@
 package com.app.data.net.services;
 
-import com.app.data.entity.restaurantSearch.RestaurantSearchEntity;
+import com.app.data.entity.restaurants.RestaurantEntity;
 
 import io.reactivex.Observable;
 import retrofit2.http.GET;
@@ -11,7 +11,7 @@ public interface RestaurantsApi {
     String RESTAURANT_SEARCH = "search";
 
     @GET(RESTAURANT_SEARCH)
-    Observable<RestaurantSearchEntity> searchResturant(
+    Observable<RestaurantEntity> getRestaurants(
             @Query("lat") String lat, @Query("lon") String lon);
 
 }
