+++ data/src/main/java/com/app/data/repository/datasource/restaurants/RestaurantDataStore.java
@@ -1,6 +1,6 @@
 package com.app.data.repository.datasource.restaurants;
 
-import com.app.data.entity.restaurantSearch.RestaurantSearchEntity;
+import com.app.data.entity.restaurants.RestaurantEntity;
 
 import io.reactivex.Observable;
 
@@ -9,5 +9,5 @@ public interface RestaurantDataStore {
     /**
      * Get an {@link Observable} which will emit restaurants nearby.
      */
-    Observable<RestaurantSearchEntity> restaurantsNearMe(String lat, String lon);
+    Observable<RestaurantEntity> restaurantsNearMe(String lat, String lon);
 }
